Move functions related to control panel to the corresponding file.
